Updated the code to run without Youtube Data API
You will not need to have an api key to run the script. I am downloading the video using pytube. I am also getting the transcript from the youtube transcript api which does not require an api key. I am getting the title of the video using beautifulsoup by scraping for the title from the browser. If you want to download the description of the video, you can uncomment the parts where you need to use the Youtube data api. The code is commented but not deleted. getVideoDescription function is still in the code but do not call it unless you uncomment the other parts and give it a file that has an api key called "API_KEY.txt"